Separate out the User logic into base Activity class.
