+++ app/src/main/java/com/sdrockstarstudios/meatheadandroid/MeatheadBaseActivity.java
@@ -1,0 +1,72 @@
+package com.sdrockstarstudios.meatheadandroid;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.firebase.ui.auth.AuthUI;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
+
+public abstract class MeatheadBaseActivity extends AppCompatActivity {
+
+    FirebaseUser currentUser;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        ActionBar actionBar = getSupportActionBar();
+        if(actionBar != null)
+            actionBar.setDisplayHomeAsUpEnabled(true);
+        currentUser = FirebaseAuth.getInstance().getCurrentUser();
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        MenuItem userProfile = menu.findItem(R.id.user);
+        if(currentUser != null){
+            userProfile.setTitle(currentUser.getDisplayName());
+        }
+        else {
+            userProfile.setTitle("");
+            userProfile.getSubMenu().removeItem(R.id.signout);
+        }
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle presses on the action bar items
+        switch (item.getItemId()) {
+            case R.id.signout:
+                signOutUser();
+                return true;
+            case android.R.id.home:
+                this.finish();
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
+    public void signOutUser() {
+        AuthUI.getInstance()
+                .signOut(this)
+                .addOnCompleteListener(new OnCompleteListener<Void>() {
+                    public void onComplete(@NonNull Task<Void> task) {
+                        Intent loginActivityIntent = new Intent(getApplicationContext(), LoginActivity.class);
+                        startActivity(loginActivityIntent);
+                    }
+                });
+    }
+
+}
